Schema combiner should set top-level format property
since the format property is a value which is switched on to
determine how a content item is rendered, it feels like its worth
strictly verifying that this value is correct.

This value is not defined in the details or links hash so we will
pass it in explicitly to the schema combiner, and in the script we can
derive it from the path of the input files.