Dummy server for example content
This is a self-contained rack app which will read example files from the
example folder.

It can be run either using the `startup.sh` shell script, or as a rack
up directly using rackup and the `config.ru` file.

We ensure that file changes propogate immediately by doing the
following:

- Set the `Cache-control: no-cache` header to ensure that the content
  store client does not cache the fetched items.
- Always read the file from disk so that changes propogate immediately.