Deprecate base_path in publishing context
We want to remove base_path from the request body in the publishing context as
it is a potential source of error. Making it optional is the first step to
removing it.

The frontend representation will continue to include `base_path`. This will be
automatically inserted by the content store.