Add Policy examples
The two different flavours of "policy", a "Policy Area" and a "Policy
Programme". There is little material difference between these two examples at
present, but they will diverge soon once we are including the link back to the
paraent "area" from the "programme". This hasn't been implemented yet, at least
not in the representation of the policy being sent to content store.