Make CI build fail on uncommitted changes to generated files
If generated files are edited manually, they get a newer timestamp than their
source files, and so don't get regenerated (unless and until their source files
are edited). This means that if a developer edits the schema.json files (which
are generated), then there could be changes in them which are not consistent
with the source files.