Add nation_applicability to Policies
This commit updates the schema for Policies to have nation
applicability. The `nation_applicability` hash contains an array of
strings which the nations that the Policy applies to along with an array
of hashes with the nation and alt_poicy_url for each nation that the
policy isn't applicable to.