Make public_updated_at required for renderable content
This is currently optional in content-store, making it required here is a route
to making it required there and backfilling the data.

We have been thinking of public_updated_at as "the updated date which users
see" and because some formats don't display this (eg gone/redirect items), it
has been optional.

However, it seems to me that it isn't about what the user sees, but when there
was last a (significant) semantic change to the content at the URL. Therefore,
everything should have a public_updated_at.

This change only adds it for "renderable content" (anything but 'gone' and
'redirect' items) because there isn't yet a gone schema and making Whitehall
produce redirect items with a (meaningful) public_updated_at is non-trivial.

As an example, it being optional meant that a placeholder without a
public_updated_at was added by publisher which then broke an expectation in
email-alert-service and stopped it from consuming any more messages from the
queue.