Support placeholder items with a variable format
We include the name of the actual format as a suffix for some placeholder items
- we want to do this with all of them, but let's do one thing at a time.

I've hardcoded this as a special case because trying to add a generic ability
to override the shared metadata.json proved too complex.

bin/combine_publisher_schema was passing through the schema name as a Pathname
object, but it should really be a string.