Add 'result_metadata_name_key' finder facet option
For rummager fields which are objects/hashes, this option
specifies which key contains the value to use for display.

Existing spec-pub finders (cma, etc), use the 'label' key
by convention, which was hard coded into finder-frontend,
but some fields used in policy finders (like organisation)
use a different key for the display value.

This option allows policy finders to fix that through the
schema, and updates example policy content to use it and
adds two other options that will be used in policy finders.

Policies are a based on the finders schema, so the change
is duplicated in both.