Update policy finder examples
Policy finders are going to display metadata for
orgs, document type and date updated.

This adds the facet options required to show them
to both the policy and program example content, so
they can be used by finder frontend as test fixtures
and for manual testing against dummy content store.