Add default order to Finders
For the Policies Finder, ordering by last update doesn't make sense as
the content being updated within each policy doesn't bubble up to update
the policy itself. This commit adds a `default_order` to the Finders and
Policy schemas along with adding an example to `/finders` of the Policy
Finder which uses this.