Fix dependencies in Makefile.
The `wildcard` call for the publisher schema was being expanded at parse
time, and therefore doesn't handle expansion of the `%` variable.  This
meant that changes to the links.json weren't being noticed, and the
corresponding schemas not being rebuild.  Fix this by using a simple
wildcard to depend on *.json in the publisher directory.

This also removes the unnecessary usage of the `%/..` to locate the
global definitons (metadata.json and frontend_links_definition.json) in
favor of pointing directly at the files.