Finder schema converter
the purpose of the finder schema converter is to convert from the
finder schema used in specialist publisher to an equivalent json-schema
describing the content format.

It produces a set of 'definitions' which will match the metadata
component of each document type. These can then be included into the
`details/metadata` sub-element of the json schema using a `oneOf`
rule[1].

This also has the advantage that it should be fairly easy to update
these definitions if the finder schemas change.

The mapping from the filename of the finder schema to the document type
name is non-deterministic. We therefore include a facility to pass a
`document_type_mapper` to the converter which can contain a list of
hard-coded rules for their conversion.

[1] http://json-schema.org/latest/json-schema-validation.html#anchor88