Correctly identify single/multi-selects
There is no way to identify whether a finder facet allows single or
multiple values in the document. The only place that this information is
encoded is in the form templates of the specialist publisher admin
interface.

This commit adds support for an injected callable which can identify
whether a field is a single or multiple select.

We can then pass in this callable with a hard-coded list of the
single/multi-select fields.