Move metadata rule to a definition
This makes it easier to temporarily swap out the `oneOf` rule for a rule
which is specific to a particular document time. You may want to do this
when trying to diagnose the cause of a validation failure relating to
invalid metadata. The json schema validator does not give a detailed
error message when failing to match a 'oneOf' rule, it just says 'none
of the requirements matched'. Temporarily swapping the rule for the
document type corresponding to the failed example will give you a more
detailed error message.