Strictly define base_path
`base_path` should always contain an absolute path. At present the
schema accepts any arbitrary string.

This definition was taken from this gist[1] which is based on RFC
3986[2].

I modified the character class slightly to move the `\-` to the end of
the character class, since I wasn't totally sure of the escaping rules
for backslash in json-schema. This makes it unambiguous (regexp
character classes will match a literal hyphen if it's the last element
of a character class, ie `[a-]` matches either `a` or `-` wheras `[a-c]`
matches `a`, `b` or `c`).

Also changed the frontend schema generator to generate a base_path which
uses the same `absolute_path` definition.

[1] https://gist.github.com/dpk/4757681
[2] https://tools.ietf.org/html/rfc3986#section-3.3