Remove change_note from specialist documents
The `change_note` field contains the same content as the last entry in
the `change_history` (the last entry is the most recent one).

It's superfluous to duplicate that information in an additional
top-level field, and doing so introduces potential for confusion and
errors.

At present it's believed that the email alert api would use the
`change_note` field when listening on the message queue, however
specialist documents are not sent to the email-alert-api in that way and
we do not intend to do that in the future.