Add optional analytics_identifier to placeholder format
We want to consistently tag pages across GOV.UK in Google Analytics with:
    * organisations
    * world_locations
    * world_organisations

The natural value to send would be the path or slug, however the limit on the
length of a dimension value in Google Analytics combined with the fact that
content is often tagged with multiple organisations/locations/etc means that
these would be truncated. Instead, we send a short code - the
analytics_identifier. Users of Google Analytics then translate these codes back
into meaningful values.

Whitehall mints these identifiers, so for content not published and rendered in
Whitehall to send this data to Google Analytics, we need to share them around.

The plan is to then include the analytics_identifier when expanding `links` so
that a content-item fetched from content-store includes the
analytics_identifiers needed to record that data in GA.