Validate publisher examples automatically
adjust to the makefile to validate publisher examples automatically.

Due to the limitations of makefiles, it's hard to have computed paths in
the dependency of a pattern rule. The only manipulation possibly is to
strip out the `%` wildcard placeholder.

To work around this, we introduced the `.frontend.valid` and
`.publisher.valid` suffixes to the validation records to allow us to
distinguish the two sets.