Add topic tags to HMRC Manuals
This change adds 'tags' under 'details', and 'topics' under those 'tags'.

These topic tags are intended to be the slugs of a Manual's topics, as
explained in the `description` property.

This is the format expected by Email Alert Service in the Current World, as
demonstrated in the Case Study schema.