Validate that redirect destinations are paths
content-store requires paths, not URLs, so the schema should too.

Querystrings and fragments are allowed in the destinations, so allow those.

One thing I've decided not to implement is that the destination is not allowed
to contain querystrings and fragments if it is a prefix redirect (ie they are
only allowed for exact redirects). The effort to implement that validation in
json-schema seems excessive, so I've ignored it.