Expose content_id in frontend representations.
This includes the content_id in the frontend representation of content
items - both in the main item, and in the expanded links.

It is becoming necessary for some of the frontend applications to know
about content_ids (eg email_alert_frontend needs the topic content_ids
in order to send them to email-slert-api to create the link to
govdelivery topics).