Allow running dummy-content-store from this repo
The behaviour of the `govuk-dummy_content_store` gem used to live in this
repo, but was split out, however, it's still most common to want to run
the dummy content store in combination with this repo.

This adds the gem as a dependncy, so it's easy to run the dummy content
store, and requires fewer repos (just the FE and the dummy content store)
which is much simpler for less technical folk to run.

I've also added some documentation to explain how to do this.

(I chatted to David Heath about this in advance, and he was happy with the
change.)