Add Procfile to run dummy content store on heroku
This will be used by a Heroku instance running the `master` version of
this repo, so there is a publically addressable version of the content-store
that people can use without having to run their own version.

This is primerily for frontend/design use, where the ideal situation is
running a single FE application, and pointing at 'live' instances of
`static` and `content-store`, which should be the only dependencies.