Add a basic schema for service manual guides
Add body, publisher and related_discussion properties. The latter is optional.

Publisher and related_discussion properties are objects containing hrefs. To
our current knowledge related_discussions always happen outside of GOV.UK
domains, so they will all be external links.

Publisher hrefs are likely to be internal URLs, but we can't create this content
at the moment so keep this field as href for now and update to content_id refs
if needed in future.