Add parent link to email_alert_signup schema
As part of the effort to convert email alerting to rely on content IDs
rather than slugs, this change adds a field to email alert signup pages
that provides a reference to the parent piece of content. This makes it
easier for the email-alert-frontend to obtain the correct content ID to
send on to the email-alert-api.