Add email_alert_type to email_alert_signup details hash
This states explicitly the kind of page a given email signup is handling.
This improves clarity when viewing an email-signup content item retrieved
from the content store, and also helps the email-alert-frontend (for example)
resolve how to format its subscription payload for the email-alert-api.

The goal of all this is to handle email subscription/alerting via content
IDs in the links hash and eventually move away from the use of slugs.