Rename email_alert_signup 'tags' property in details hash
This changes the name of this property from 'tags' to 'signup_tags'. The
latter more accurately reflects the semantics of what this property is
supposed to represent. We aren't really stating that an email alert page
is tagged to (grouped under) a policy/topic, in the same way that other
content might be. Rather, these alert pages have their own, reasonably
well-defined purpose - to provide all the data necessary to create
a subscription to alerts for a given piece of content.

The result of this change is that email-alert-frontend can conveniently
send subscriptions to the email-alert-api with the correct tags, while
email-alert-service can continue to provide email-alert-api with the
'tags' hash of published items.