Enforce presence of signup_tags in details, remove tags
This is a followup to 1bc4952, which provides detail around why this
change was introduced in the first place. Also note that this rename
fixes a subtle bug - when this field was still called 'tags', published
changes to the email_alert_signup content item would result in an email
alert. This was not the intended behaviour.