Use links hash to expand content ids from details hash
Now that we've started referencing content using content ids, we need a way
to expand them so that front ends could render links to the referenced
content items. Publishing team is planning to build a dependency resolution
service, but it's not ready yet. Temporarily we can replicate all content
ids from the details hash in the links hash, as all content ids in the links
hash are expanded in content store responses.