Add contact front-end example
We add `organisations` to the links in order to be able to link to content-store organisations in future. We will republish all contact content items with this format in a PR for `contacts-admin`.

At the moment the `contacts-admin` publisher app still imports `organisations` from `whitehall` and sends a contact's `organisation` data in the details hash of a content item. This will probably be changed in future, but for the moment we chose to support both approaches (so having `organisation` data both in the details hash as well as the links hash). The reason why we cannot not just switch to using the `organisations` from the `content-store` right now is that in `contacts-frontend` we require the `organisation.abbreviation` attribute, which is currently not stored for `content-store` organisations.