Port combined_publisher_schemas to Rake
Use Rake to resolve dependencies.

Creates separate Rake FileLists for the 3 types of schema we generate
(v1, v2 details and v2 links), with Rake Rules which specifies which
source files to use. Calls the `SchemaCombiner` class to merge the JSON
files. Some formats are excluded as they’re hand generated, these will
be handled in a future commit.

Replaces Make rule these files with a call to the Rake task.