Create handcrafted v2 schema
redirect, gone and special_route aren’t created with the normal
workflow, and until now have just used the same schema for v2 as v1.
This is invalid as v2 doesn’t allow content_id or update_type in the
published document.

This commit handcrafts the v2 schema for these types, and updates the
Rake task to copy them directly into dust.

content_id and update_type are still valid but not required for ease
of upgrading. Publishing API will ignore them. When apps are updated
to not send these properties, we can remove them from the schemas
entirely.