Add primary_parent definition to Travel Advice...
Adds a simple breadcrumb structure similar to the current ancestry tree
currently found in section tags on GOV.UK This is a tentative first pass
at how breadcrumbs are defined. Initially these are required as part of
the details for travel advice while the breadcrumb data is static, this
is expected to be moved out into the metadata top level once dependency
resolution has been implemented.