Add update_type to V2 metadata schema
As Applications move to V2 they will start to the use the Publishing API
as a data store. When they do this, in order to persist `update_type`
between draftings, it will need to be stored as part of the metadata.
This commit adds it to the V2 metadata but doesn't make it required, as
the `update_type` can also be provided when calling the `publish`
command on the content item.