Remove details>organisations from required attributes in hmrc manuals
The hmrc-manuals-api is being updated to publish organisation data in the
links hash rather than the details hash, as per current convention for
documents across GOVUK. This change drops details>organisations as a
required attribute. Cleanup of details>organisations in the examples
will follow after further updates to the hmrc-manuals-api.