Move Service manual guide content owner from details to links
Now that we have community pages to reference, we can replace this attribute
with a standard content_id type link that will get expanded and will carry
the current value of title and URL in it.

We'll need to re-PUT all existing SM guide drafts (there are not published items
at the moment) and remove content owner field from the details hash once it's
done.

I've made the details content owner an optional field. Once migrated we can
make the content_owners field in links mandatory if it makes sense.

Also, frontend will need to be updated to prefer the links field and fall back
to the one in details.

As an extra this adds optional topics link to the example files.