Add fields for topical event placeholders
We can't migrate topical events yet, because they are complex.

Topical event about pages can be migrated, but they need the end_date of the
topical event they are a part of in order to render the page.

The examples provide the topical events for the examples of topical event about
pages.