Expose ruby version via Gemfile
Not all environments will pick up a `.ruby-version` file when deciding
which version of ruby to use. Heroku uses a Gemfile's `ruby x.y.z`

This repo is deployed to Heroku as an instance of `dummy_content_store`
serving it's examples [1].

In order for this app to be deployed to Heroky at the right ruby version
the version is exposed via the Gemfile as well, but not duplicated, so
there isn't a risk of the versions being set independently and falling
out of sync.

When a version isn't set, Heroku defaults to `2.0.0`, which it is sun
setting [2], so a Heroku-recognised version needs to be exposed for
the instance of the app to continue working.

This also makes the ruby version specific to the patch level, not just
minor. The difference in version is important when running as an
application, eg, the dummy content store.

Although this repo isn't an application intended to be deployed, it's
heroku instance is a supported workflow [3] and it's often run locally
during development, when not working against a live content store, and
in both these cases we should be pinned to a specific version of ruby.

Exposing the `.ruby-version` via the `Gemfile` _is_ a little ugly, but
it's a pattern we've used before on apps we want to run in heroku [4]
and as our apps become simpler, and easier to run independantly of the
rest of our stack, this is a pattern we may want to follow on more of
our apps.

[1]: https://govuk-content-store-examples.herokuapp.com
[2]: https://devcenter.heroku.com/articles/ruby-support#ruby-versions
[3]: https://github.com/alphagov/govuk-content-schemas/blob/master/docs/running-frontend-against-examples.md
[4]: https://github.com/alphagov/government-frontend/commit/98c05c01198133c8afd88a3ad4bb0a9cd06d9ba0#diff-8b7db4d5cc4b8f6dc8feb7030baa2478R3