Allow additional properties in front-end links
For topical event about pages we want some extra information - the
topical event start and end date. This will allow the front-end to show
whether the topical event is archived or not.

Rather than including the specific details and start/end dates in the
properties, we are allowing additionalProperties. Per format link
definitions are hard to define at the moment, for instance start/end
date may have different meaning based on format.

This is a short term solution in lieu of more programatically generated
frontend links schemas. When the new pub api dependency work is completed
the schemas will be used to drive what fields are included so it'll be
able to be embedded correctly in the frontend schema.