A new schema service_manual_topic.
The new service manual is going to get it's own topic format, not for a specific technical requirement, but to help keep it better conceptually grouped together with the other service manual formats and separate from the non-service manual formats in government-frontend. Once the formats have settled and found their rightful frontend we hope to be able to delete this service_manual_topic format.

This schema was based on topic but with the following notable changes:

 * details->internal_name removed from the schema
 * details->beta removed from the schema
 * details->groups removed from the schema
 * the publisher_v2/examples directory removed