Add document_type and schema_name to top level.
These are alternatives to format; it is an error to supply both sets. To
enable this, the top-level properties has been replaced with a 'oneOf'
matcher with two choices, each of which includes all the existing fields plus
either format or the two new fields. This is because JSON Schema does not
support extending a schema with alternative options, only replacing it.

To reduce duplication in the generated schemas, the links and details that
would have been included in the top level fields have been moved to the
definitions hash with only a reference included in the two options.