Move content_owner to links for service_manual_guide
The pages which represent communities who curate guides now exist. Before we had a fixed href and title for the reference from the guide to the relevant community but now we can use content_ids.

We keep the content_owner defintion while they still exist in the content store but they are no longer required. Once all items have been republished we can lose the definition.