Remove double space in HTML publication example
This is causing a test failure in the frontend intergration test, where
the two spaces are collaposed to one by Capybara (we think), causing the
two strings to be different.

We're looking at a longer term fix that can ignore to ignore the whitespace
differences as part of the tests, so we don't have to be so careful in
the individual examples.

But in the meantime fix this example to unblock HTML Publication work.