Update placeholder format logo.crest property.
Whitehall is being updated to stop publishing crests with the `custom`
and `no-identity` values, and instead will publish `null`.

This change updates the placeholder format to be compatible, and also
one of the html_publication examples to conform.

This will fix an issue in `government-frontend`, which is displaying the
"Environment Agency" crest incorrectly.

Relevant Trello ticket:
https://trello.com/c/cJGrtnc1/286-5-html-publications-migration-front-end-work-large