Add government and political definitions
The plan is to allow government to be a link, so that `current`, names
and slugs can be updated automatically without having to republish.
However governments represent URL-less content, which the content store
doesn’t currently handle.

As an intermediate to unblock detailed guides, put government into the
details hash instead.

From `government` and `political`, whether the history banner should
show can be derived.