A boolean to apply the accordion for service_manual_topic
The service_manual_topic has multiple sections which are normally collapsed. Sometimes there aren't enough sections or items within a section to justify to collapsing and it appears there aren't a set of rules that can be applied which will let the frontend decide for itself. Instead we are adding this boolean which the publisher will set so a pair of human eyes can decide whether it should be collapsed or expanded.

If the attribute is false or missing then the default behaviour will be that it is expanded.