Add documents to publication format
Documents is an array of attachments, each attachment contains the
markup necessary to render it within a govspeak block.

This is an intentionally dumb approach, and matches the approach taken
for embedded attachments.

Use an array so that attachments can be ordered, counted, etc.

Details on attachments decisions:
https://trello.com/c/HLb48j22/232-govspeak-implementing-the-ui-of-attach
ments-medium
https://docs.google.com/document/d/1Flx_eCUzixmkTT6WgjCN2bJzQ8HuVCwMga5R
xUsv07M/edit

"Keep attachment template in WH Publisher, rendering it and sending
attachment HTML to the content store. Copy the CSS/JS for attachments
to the Govspeak Component, as we did in the past for embedded contacts.

This solution is required for embedded attachments, but can also be
used to migrate attachment lists (used on formats like Publications),
however should changes to the attachment template be required, all
editions with embedded attachments would need re-publishing."

Long term solution:
https://gov-uk.atlassian.net/wiki/display/GOVUK/RFC+39%3A+Embedded+relat
ional+content+in+govspeak