Update examples...
In order to transition to having the  as a top level property,
we need backwards compatibility with formats already expecting it in the details
hash eg. case studies. So these examples contain duplicate properties for now,
these should be removed once the frontends are using the top level withdrawn_notice.