Add top & second level browse pages to details hash
We are adding top level and second level browse pages into the details
hash because we need them to be able to be ordered when the content is
curated in collections publisher.

In the past we made a decision of not preserving the order for links,
which can be found here: https://gov-uk.atlassian.net/wiki/display/FS/RFC+36%3A+Stop+preserving+order+for+links

One of the key reasons for this decision was to make sure presentation
concerns don't get put in the links hash. That's why we are using details instead: neither of
these fields are needed to describe how the browse pages link together
(we have active_top_level for that).

There will be following commits to remove these two fields from the links
hash once we have fixed the frontend apps.

Trello:
https://trello.com/c/KZJigx96/676-ordering-of-browse-pages-no-longer-respected

@davidslv & @MatMoore