Provide a way to use format specific metadata...
Some exceptional formats will be pathless, for example contacts from Whitehall
do not have a base_path or any routes. These formats are constrained in the publishing
API and this should be reflected in the content schemas. So provide a way to override
the common metadata on a per-format basis.