Add source for schema for the message queue payload
The publishing api sends JSON documents out on a "message queue". These JSON
documents include a representation of a content item. This new schema aims to
describe these JSON documents, as they are generated currently by the
publishing api.

The schema should accept any current output of the publishing api, but is
incomplete in many areas (e.g. loose validation for expanded_links), and so is
not very useful for consumers of the message queue in terms of documentation of
what can be expected.

Originally I did set out to make a schema per format as is done for the
publisher and frontend schemas, but ran in to issues with the behavior of the
schema combiner, particularly with regard to the mutual exclusivity of the
format and document_type + schema_name fields as the message queue payload
includes all three. As a result of this, I switched to building a schema to
match the message queue payload, for any format.