Add 'temporary_update_type' to specialist documents
When adding an attachment to a document, it sends the
document to the Publishing API with the attachment's
metadata. If the user hasn't chosen an update_type for
the document, we intend to set it to 'minor' and set
this 'temporary_update_type' field to true. We can then
present a choice of update types to the user rather
than pre-selecting 'minor' for them.

Relates to:
https://trello.com/c/oxxALgmy/187-changes-to-update-type-for-raib-reports