Allow email alert signups to work with links other than countries
The email-alert-api uses the links hash within subscriber_lists as a matcher, examining new documents that get published for matching `links` and `document` types and sending emails to relevant lists.

The schema currently limits the contents of the links hash to only permit countries, which means that only links to `countries` can be subscribed to.

We are building email alerts into the service manual, allowing users to subscribe to service manual topics. Therefore we need to create email_alert_signups based on documents that link to `service_manual_topics`, so we need to allow for this in the schema.

Rather than adding `service_manual_topics` as another 'special case' this instead allows exactly one property with a alphabetical, underscored key to be provided as long as the value is an array.