Remove `format` from allowed attributes in schemas
Currently we allow two types of schemas: one with `format` and one with
`document_type` and `schema_name`. We've done the work to remove usage
of `format` so we can disable the first format.

This doesn't completely remove the `oneOf` clause from the schemas
because that is more involved (the tests rely heavily on this
behaviour). This is the minimum we can do now to disallow `format` in
publisher schemas. Follow up will remove the `oneOf` thing.