Remove children/*/parent from service standard example
Currently it doesn't seem possible to have a frontend example of a parent/child relationship which has been link-expanded that matches it's frontend schema.

It seems the definitions of each link type are overwritten for the frontend schemas so that they are "#/definitions/frontend_links". https://github.com/alphagov/govuk-content-schemas/blob/b283419dc1f43d69396e88fdd559ea2f4ce1889b/lib/govuk_content_schemas/frontend_schema_generator.rb#L81

The "#/definitions/frontend_links" definition mostly matches an expanded link but falls down when it gets into the links of the expanded link where it expects every link type to be a "#/definitions/guid_list".

Here are a couple of heavily redacted examples to show what we have and what the frontend schema wants. The difference is that the former has it's /links/children/*/links/parent expanded whereas the later has a guid list.

This is the structure of a link expanded content item in the wild:
```
{
  "base_path": "/service-manual/service-standard",
  "links": {
    "children": [
      {
        "base_path": "/service-manual/service-standard/have-a-multidisciplinary-team",
        "links": {
          "parent": [
            {
              "base_path": "/service-manual/service-standard",
              "links": {}
            }
          ]
        }
      }
    ]
  }
}
```

This is the structure of a content item that the schema expects:
```
{
  "base_path": "/service-manual/service-standard",
  "links": {
    "children": [
      {
        "base_path": "/service-manual/service-standard/have-a-multidisciplinary-team",
        "links": {
          "parent": ["00f693d4-866a-4fe6-a8d6-09cd7db8980b"]
        }
      }
    ]
  }
}
```

I'm not sure how much this matters? For parent/child link expanded stuff, the /links/children/*/links/parent data is also available in the root object.. it's a link expanded reference to itself. I think a slight problem is that our examples won't look exactly like they do in the wild. A larger potential problem is for another link type where the relationship isn't self-referential like the parent/child one and we actually want that data?