Associate topics using parent-child relationship
This is a bidirectional relationship which means we don’t need to republish the homepage every time we add or remove a topic.

We could configure our own bidirectional expansion in the publishing api, but this keep things simple until such a time when we have e.g. multiple types of children.