Don't allow additional properties for frontend links
This changes the way the links are validated in frontend schemas.

Since #232 the frontend links have additionalProperties set to true. I
think this is undesirable because it allows all data in the frontend
links, defeating the purpose of schemas.

- Two detailed guide examples have the links incorrectly nested (links
-> parent -> parent instead of links -> parent -> links -> parent).
This wasn't picked up by the validation. This could have caused
developers to build pages that work only with the incorrect example.
- The service manual example included a expanded_links example, which
is not used anymore (and should never have been used in frontends).